walkgroupFindExitPointFn_800dc398 81.07->81.18: pairId u16->int (full-width pairId compare, mask moves to store; #58)

+++ src/main/dll/dll_0014_unk.c
@@ -2294,7 +2294,7 @@ void walkgroupFindExitPointFn_800dc398(void)
     u8 gb;
     u8 e;
     u32 j2;
-    u16 pairId;
+    int pairId;
     u8 found;
     int searchCount;
     ObjfsaWalkGroup* wg;
